Updated lex & parse examples + others
